restructured logic and functions into seperate files
