+++ js/main.js
@@ -1,28 +1,10 @@
-import {createCard, handleEditBtn, handleDoneBtn, handleDeleteBtn} from "./card.js"
+import { handleCardClick } from "./card.js"
+import { handleFormSubmit } from "./form.js";
 
 document.getElementById("destination_form").addEventListener("submit", (e) => {
-    debugger;
-    e.preventDefault();
-
-    const destination = getInputValue("input-destination-name");
-    const location = getInputValue("input-location");
-    const photoUrl = getInputValue("input-photo-url");
-    const description = getInputValue("input-description");
-
-    const newCard = createCard(destination, location, photoUrl, description);
-
-    document.getElementById("cards_container").prepend(newCard);
-
-    document.getElementById("destination_form").reset();
+    handleFormSubmit(e);
 });
 
 document.getElementById("cards_container").addEventListener("click", (e) => {
-    e.target.classList.contains("delete-btn") ? handleDeleteBtn(e)
-    : e.target.classList.contains("edit-btn") ? handleEditBtn(e)
-    : e.target.classList.contains("done-btn") ? handleDoneBtn(e)
-    : {}
+    handleCardClick(e);
 });
-
-function getInputValue(id) {
-    return document.getElementById(id).value;
-}
